pyproject.toml
[project]
name = "video-splitter"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.4",
    "python-dotenv>=1.0.1",
    "pytubefix>=8.3.0",
    "sqlmodel>=0.0.22",
]
